Finish initial working version of more flexible dynamic menubar
Initial version of the dynamic datatables menubar def was focused on being simply a container for filters. That ended up not working though, because we kept getting requests to place other content in the menubar. Now the implementation is a much more flexible model that can be configured more from the json definition. There is more freedom to not follow the recommended layout, but another filter can be added for that later at a higher level.